requirements.txt
blinker==1.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
DateTime==5.5
Flask==3.0.3
Flask-Login==0.6.3
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
oauthlib==3.2.2
pytz==2024.2
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.0.3
Werkzeug==3.1.3
zope.interface==7.0.3
